Fixed codefactor issues
