+++ src/Downloader.Test/UnitTests/ChunkDownloaderTest.cs
@@ -121,40 +121,4 @@ namespace Downloader.Test.UnitTests
             Assert.ThrowsExceptionAsync<OperationCanceledException>(CallReadStream);
         }
     }
-
-    [TestClass]
-    public class ChunkDownloaderOnMemoryTest : ChunkDownloaderTest
-    {
-        [TestInitialize]
-        public override void InitialTest()
-        {
-            _configuration = new DownloadConfiguration {
-                BufferBlockSize = 1024,
-                ChunkCount = 16,
-                ParallelDownload = true,
-                MaxTryAgainOnFailover = 100,
-                Timeout = 100,
-                OnTheFlyDownload = true
-            };
-            _storage = new MemoryStorage();
-        }
-    }
-
-    [TestClass]
-    public class ChunkDownloaderOnFileTest : ChunkDownloaderTest
-    {
-        [TestInitialize]
-        public override void InitialTest()
-        {
-            _configuration = new DownloadConfiguration {
-                BufferBlockSize = 1024,
-                ChunkCount = 16,
-                ParallelDownload = true,
-                MaxTryAgainOnFailover = 100,
-                Timeout = 100,
-                OnTheFlyDownload = true
-            };
-            _storage = new FileStorage();
-        }
-    }
 }
